fix: changed type of price to be BigDecimal

+++ src/main/java/com/podzilla/courier/dtos/delivery_tasks/DeliveryTaskResponseDto.java
@@ -3,7 +3,9 @@ package com.podzilla.courier.dtos.delivery_tasks;
 
 import com.podzilla.courier.models.DeliveryStatus;
 
-public record DeliveryTaskResponseDto(String id, String orderId, String courierId, Double price, DeliveryStatus status,
+import java.math.BigDecimal;
+
+public record DeliveryTaskResponseDto(String id, String orderId, String courierId, BigDecimal price, DeliveryStatus status,
                                       Double orderLatitude, Double orderLongitude, Double courierLatitude,
                                       Double courierLongitude) {
 }
